Project Matching and Course Creation

+++ app.py
@@ -7,6 +7,7 @@ from hashlib import scrypt
 import secrets
 import json
 from os import environ, path
+from random import shuffle
 app = Flask(__name__, static_folder="frontend/build")
 
 
@@ -152,6 +153,36 @@ def getProfileInfo():
         conn.commit()
     return out
 
+@app.route("/api/createcourse", methods=["POST"])
+def createcourse():
+    name = request.json.get("name", None)
+    description = request.json.get("description", None)
+    term = request.json.get("semester", None)
+    crn = request.json.get("crn", None)
+    groupsizemax = request.json.get("groupsizemax", None)
+    groupsizemin = request.json.get("groupsizemin", None)
+    maxrankings = request.json.get("maxrankings", None)
+
+    s_username = session.get('username', None)
+
+    if not (name and term and crn):
+        return jsonify({"success": False, "error": "You must give a name, term, and course code."})
+
+    with conn.cursor(cursor_factory=RealDictCursor) as c:
+        c.execute(("SELECT university_id FROM account WHERE username = %(username)s"), {"username": s_username})
+        university_id = c.fetchone()["university_id"]
+        c.execute(("INSERT INTO course (crn, term, title, university_id, groupsizemax, groupsizemin, maxrankings) "
+                    "VALUES (%s, %s, %s, %s, %s, %s, %s) "
+                    "RETURNING *"), (crn, term, name, university_id, groupsizemax, groupsizemin, maxrankings))
+        cid = c.fetchone()["id"]
+        print (cid)
+        print (s_username)
+
+        c.execute(("INSERT INTO instructor (account_id, course_id) VALUES ((SELECT id FROM account WHERE username = %(username)s), %(cid)s)"), {"username": s_username, "cid": cid})
+        out = jsonify({"success": True})
+        conn.commit()
+    return out
+
 @app.route("/api/createproject", methods=["POST"])
 def createproject():
     name = request.json.get("name", None)
@@ -198,7 +229,6 @@ def editProject():
         conn.commit()
     return out
 
-#this should probably do something is this project is a foreign key for something other than tags...
 @app.route("/api/deleteproject", methods=["POST"])
 def deleteproject():
     id = request.json.get("id", None)
@@ -207,6 +237,16 @@ def deleteproject():
         return jsonify({"success": False, "error": "You must specify project id"})
 
     with conn.cursor() as c:
+        c.execute("SELECT * FROM preference WHERE project_id = %s", (id, ))
+        preference = c.fetchone()
+        c.execute("SELECT * FROM approved WHERE project_id = %s", (id, ))
+        approved = c.fetchone()
+        c.execute("SELECT * FROM project_group WHERE project_id = %s", (id, ))
+        group = c.fetchone()
+
+        if preference or approved or group:
+            return jsonify({"success": False, "error": "This project has an attached group or preference that prevent deletion"})
+
         c.execute("DELETE FROM project_tags WHERE project_id = %s", (id, ))
         c.execute("DELETE FROM project WHERE id = %s", (id, ))
         out = jsonify({"success": True})
@@ -238,6 +278,20 @@ def listOrganizations():
         conn.commit()
 
     return jsonify({"success": True, "organizations": key_val})
+
+# this is a copy paste job and needs to be improved to detail crn, term, instructor, and other distinguishing features
+@app.route("/api/listCourses", methods=["GET"])
+def listCourses():
+    with conn.cursor(cursor_factory=RealDictCursor) as c:
+
+        c.execute("SELECT * FROM course")
+        key_val = {}
+        coursesToShow = list(c)
+        for i in coursesToShow:
+            key_val[i['id']] = i['title']
+        conn.commit()
+
+    return jsonify({"success": True, "courses": key_val})
 
 @app.route("/api/search", methods=["GET"])
 def search():
@@ -353,6 +407,23 @@ def getCoursesToDrop():
         coursesToShow = list(c)
     return jsonify({"success": True, "courses": coursesToShow})
 
+@app.route("/api/courses_of_instructors", methods=["GET"])
+def getCoursesOfInstructors():
+    username = session.get('username', None)
+    with conn.cursor(cursor_factory=RealDictCursor) as c:
+        c.execute(("WITH my_acct AS ("
+                "SELECT id, university_id "
+                "FROM account "
+                "WHERE username = %(username)s"
+                ") "
+                "SELECT course.id, course.crn, course.university_id, course.term, course.title, course.groupsizemin "
+                "FROM course "
+                "INNER JOIN my_acct ON my_acct.university_id = course.university_id "
+                "INNER JOIN instructor ON (instructor.course_id = course.id AND instructor.account_id = my_acct.id) "), {"username": username})
+        coursesToShow = list(c)
+
+    return jsonify({"success": True, "courses": coursesToShow})
+
 @app.route("/api/courses_to_add", methods=["GET"])
 def getCoursesToAdd():
     username = session.get('username', None)
@@ -383,23 +454,23 @@ def addCourse():
                 "FROM account "
                 "WHERE username = %(username)s"
                 ), {"username": username})
-        
+
         d = c.fetchone()
         acct_id = d["id"]
         uid = d["university_id"]
 
         c.execute("SELECT * "
                 "FROM enroll "
-                "WHERE account_id = %(acct_id)s " 
+                "WHERE account_id = %(acct_id)s "
                 "AND course_id = %(cid)s "
-                "AND university_id = %(uid)s", 
+                "AND university_id = %(uid)s",
                 {"acct_id":acct_id, "cid":cid, "uid":uid})
 
         if not (c.fetchone()):
             c.execute("INSERT INTO enroll (account_id, course_id, university_id) VALUES (%(acct_id)s, %(cid)s, %(uid)s)", {"acct_id":acct_id, "cid":cid, "uid":uid})
             conn.commit()
             return jsonify({"success": True})
-            
+
     return jsonify({"success": False})
 
 @app.route("/api/drop_course", methods=["POST"])
@@ -411,14 +482,14 @@ def dropCourse():
                 "FROM account "
                 "WHERE username = %(username)s"
                 ), {"username": username})
-        
+
         d = c.fetchone()
         acct_id = d["id"]
         uid = d["university_id"]
 
         c.execute("DELETE FROM enroll WHERE ((account_id = %(acct_id)s) AND (course_id = %(cid)s) AND (university_id = %(uid)s))", {"acct_id":acct_id, "cid":cid, "uid":uid})
         conn.commit()
-            
+
     return jsonify({"success": True})
 
 @app.route("/api/project/<int:id>", methods=["GET"])
@@ -465,6 +536,18 @@ def getProjectPreferences():
     with conn:
         with conn.cursor(cursor_factory=RealDictCursor) as c:
             c.execute("SELECT project.* FROM preference INNER JOIN project ON preference.project_id = project.id INNER JOIN account a ON a.id = preference.account_id WHERE a.username = %s ORDER BY preference.ranking ASC", (session["username"], ))
+            projects = c.fetchall()
+            return jsonify({"success": True, "projects": projects})
+
+@app.route("/api/project/studentassigned", methods=["GET"])
+def getProjectsAssignedToStudents():
+    if not "username" in session:
+        return error("You must be logged in!")
+    with conn:
+        with conn.cursor(cursor_factory=RealDictCursor) as c:
+            c.execute("SELECT DISTINCT project.* FROM project_group INNER JOIN project ON project_group.project_id = project.id "
+                "INNER JOIN member ON project_group.id = member.project_group_id INNER JOIN account a "
+                "ON a.id = member.account_id WHERE a.username = %s", (session["username"], ))
             projects = c.fetchall()
             return jsonify({"success": True, "projects": projects})
 
@@ -517,6 +600,26 @@ def preferenceProject():
                       (session["username"], projectId))
             (newRank, ) = c.fetchone()
             return jsonify({"success": True, "newRank": newRank})
+
+@app.route("/api/group/studentGroupListing/<int:courseid>", methods=["GET"])
+def getGroupListing(courseid):
+    with conn.cursor(cursor_factory=RealDictCursor) as c:
+        c.execute("SELECT account.name AS studentname, account.id AS studentid, project.name AS projectname, project.id AS projectid "
+        "FROM project, member, project_group, account "
+        "WHERE (project_group.course_id = %s AND project_group.project_id = project.id AND member.project_group_id = project_group.id AND member.account_id = account.id)", (courseid, ))
+        students = list(c)
+
+        return jsonify({"success": True, "listing": students})
+
+@app.route("/api/group/runGroupMatch", methods=["POST"])
+def runGroupMatch():
+    courseid = request.json.get("courseid", None)
+    success = createGroups(courseid)
+
+    if success:
+        return jsonify({"success": True})
+    else:
+        return jsonify({"success": False, "error": "Something went wrong"})
 
 def error(s):
     return jsonify({"success": False, "error": s})
@@ -581,3 +684,173 @@ def getProjectById(id):
             else:
                 project["ranking"] = None
             return project
+
+# calculates single group satisfaction score based on formula in Freiheit & Wood
+# arguments: dictionary of students to preferences, id of group, members of group, max allowed rankings
+# return: score
+def calculateSatisfaction(prefs, projectid, groupmembers, maxrankings):
+    maxpossible = len(groupmembers) * (maxrankings + 1)
+
+    score = 0
+    for m in groupmembers:
+        if projectid in prefs.get(m):
+            rank = prefs.get(m).index(projectid) + 1
+            score += ((((maxrankings - rank) + 1) * (maxrankings - rank)) / 2) + 1
+        else:
+            score += (maxrankings / 2) + 1
+
+    score = score / maxpossible
+    return score
+
+# matches groups for a course.
+# arguments: a list of students, dictionary of students to preferences,
+    # minimum group size, maximum group size, maximum rankings per student
+# return values: a dictionary of projects to a list of assigned students, average satisfaction score
+# this will modify the student list, but as we're just randomizing every time it doesn't really matter
+# TODO: HEAVY EDGE CASE TESTING. Some cases have been noted in TODOs throughout this function. Others probably have yet to be found.
+# This could also be cleaned up a lot.
+def matchGroups(students, prefs, groupsizemin, groupsizemax, maxrankings):
+    # create dictionary of project ids to an empty list
+    # I know there's probably a cleaner way to make this dictionary, but I couldn't figure it out
+    groups = {}
+    for val in prefs.values():
+        for pr in val:
+            if not pr in groups:
+                groups[pr] = []
+
+    # create a shuffled copy of students who are not yet assigned
+    unassigned = students.copy()
+    shuffle(unassigned)
+
+    # initial loop: assign choice if possible, (assigning all to 1st, then to 2nd, etc...)
+    # if not possible (i.e. group is full) keep in unassigned
+    for m in range (0, maxrankings):
+        for i in range (0, len(unassigned)):
+            if len(prefs.get(unassigned[i])) > m:
+                if len(groups.get(prefs.get(unassigned[i])[m])) < groupsizemax:
+                    # there is space, assign student
+                    groups.get(prefs.get(unassigned[i])[m]).append(unassigned[i])
+                    unassigned[i] = -1 # turns out deleting from python lists while iterating sucks, this is my solution
+        unassigned = list(filter(lambda x: x != -1, unassigned))
+
+    # calculate number of remaining spots
+    spotsRemaining = 0
+    for members in groups.values():
+        spotsRemaining += (groupsizemax - len(members))
+
+    # TODO: honestly this is a really big bug that I don't know how to deal with.
+    # basically means that everyone preferenced the same projects.
+    # This should be rare, but needs to be dealt with somehow.
+    # Right now, this will just assign extra students to a non-existent project, give this run a low score, and bail
+    if spotsRemaining < len(unassigned):
+        groups[-1] = unassigned
+        score = 0
+        return groups, score
+
+    # remove underassigned projects, starting with smallest groups and making sure there are enough spots
+    # TODO: order by popularity. Since popularity is not implemented, this could delete projects that maybe shouldn't be.
+    if spotsRemaining > len(unassigned):
+        # don't sue me for this
+        toRemove = []
+        for i in range(0, groupsizemax):
+            if (spotsRemaining - groupsizemax) < len(unassigned):
+                break
+            else:
+                groupsOfSize = {k: v for k, v in groups.items() if len(v) == i}
+                for g in groupsOfSize:
+                    if (spotsRemaining - groupsizemin) < len(unassigned):
+                        break
+                    else:
+                        toRemove.append(g)
+                        for s in groups.get(g):
+                            # TODO: attempt to put student in a different choice
+                            unassigned.append(s)
+
+        for r in toRemove:
+            if r in groups:
+                del groups[r]
+
+    # start the bumping routine with unassigned students
+    # this should also take care of the case where there are full groups,
+    # but to fit all students properly none should be full (such as 6 students to 2 groups of 3-4 members)
+    # for u in unassigned:
+        # TODO
+
+    # randomly assign any remaining unassigned to fill groups
+    # there might be a more student friendly way to do this, but that's going to take some thinking
+    shuffle(unassigned)
+
+    # fill groups to groupsizemin
+    underassignedGroups = {k: v for k, v in groups.items() if len(v) < groupsizemin}
+    for ug in underassignedGroups:
+        spots = groupsizemin - len(groups.get(ug))
+        for i in range(0, spots):
+            if len(unassigned) > 0:
+                groups.get(ug).append(unassigned.pop())
+            else:
+                break
+        if len(unassigned) == 0:
+            break
+
+    # if there are still unassigned students, fill to groupsizemax
+    # this makes sure no groups are huge while others are small
+    if len(unassigned) > 0:
+        while len(unassigned) > 0:
+            openGroups = {k: v for k, v in groups.items() if len(v) < groupsizemax}
+            for og in openGroups:
+                if len(unassigned) > 0:
+                    groups.get(og).append(unassigned.pop())
+                else:
+                    break
+
+    # calculate average satisfaction score
+    # if group is somehow still too small, give it a bad score
+    score = 0
+    for g in groups:
+        if (len(groups.get(g)) >= groupsizemin):
+            score += calculateSatisfaction(prefs, g, groups.get(g), maxrankings)
+    score = (score / len(groups))
+
+    return groups, score
+
+def createGroups(course_id):
+    with conn:
+        with conn.cursor() as c:
+            c.execute("SELECT groupsizemin, groupsizemax, maxrankings FROM course WHERE id = %s", (course_id, ))
+            (groupsizemin, groupsizemax, maxrankings) = c.fetchone()
+
+            if not groupsizemin or not groupsizemax or not maxrankings:
+                return False
+
+            # list of students (the distinct probably isn't strictly necessary, but lets just make sure)
+            c.execute("SELECT DISTINCT account_id FROM enroll WHERE course_id = %s", (course_id, ))
+            students = list(c)
+
+            if len(students) == 0:
+                return False
+
+            # dictionary of students to preferenced projects
+            # the keys are student ids, with the value containing an ordered list
+            # of their prefered project ids, from first to last choice
+            prefs = {}
+            for s in students:
+                c.execute("SELECT project_id FROM preference WHERE account_id = %s ORDER BY ranking ASC", (s, ))
+                prefs[s] = list(c)
+
+            # run matching algorithm to find best satisfaction
+            bestscore = 0
+            finalgroups = {}
+            for i in range (0, 30):
+                (groups, score) = matchGroups(students, prefs, groupsizemin, groupsizemax, maxrankings)
+
+                if score > bestscore:
+                    finalgroups = groups
+
+            # insert finalized groups
+            for g in finalgroups:
+                c.execute("INSERT INTO project_group (project_id, course_id) VALUES (%s, %s) RETURNING id", (g, course_id))
+                (projectgroupid, ) = c.fetchone()
+                for s in finalgroups.get(g):
+                    c.execute("INSERT INTO member (account_id, project_group_id) VALUES (%s, %s)", (s, projectgroupid))
+
+    return True
